fix: Rename main package git issue
